Craetion of Public and private key!!
This file contains information on how to create a private and public key for the RSA cryptosystem and how to decode a message. You can also test it with your own text  but it contains some sample examples. If you want to give me an idea, reachout to me at pierre.ntakirutimana@aims.ac.rw.